Backend/requirements.txt
fastapi
uvicorn 
pandas
numpy
scikit-learn
youtube-transcript-api 
google-generativeai 
json5 
python-dotenv
redis
grok